+++ packages/wagmi/src/wallets/okx-wallet.tsx
@@ -1,0 +1,11 @@
+import { metadata_OkxWallet } from '@ant-design/web3-assets';
+
+import { EthereumWallet } from '../interface';
+import { UniversalWallet } from './universal-wallet';
+
+export const OkxWallet: EthereumWallet = (metadata) => {
+  return new UniversalWallet({
+    ...metadata_OkxWallet,
+    ...metadata,
+  });
+};